style/feat/fix: removed more blank lines, fixed sobre syntax and typing problems, moved css from variables to html element to better syntax and manuting
