+++ package.json
@@ -18,14 +18,17 @@
     "axios": "^1.6.8",
     "clsx": "^2.1.0",
     "dotenv": "^16.4.5",
+    "i": "^0.3.7",
     "install": "^0.13.0",
     "lucide-react": "^0.372.0",
     "next": "14.1.3",
     "next-google-fonts": "^2.2.0",
+    "npm": "^10.8.0",
     "qrcode.react": "^3.1.0",
     "react": "^18",
     "react-dom": "^18",
     "react-icons": "^5.0.1",
+    "react-toastify": "^10.0.5",
     "swiper": "^11.0.7",
     "tailwind-merge": "^2.3.0"
   },
