+++ src/components/Home/heroBanner/index.tsx
@@ -1,0 +1,20 @@
+"use client"
+
+// exportaçoes de componentes
+import { HomeDesktop } from "../desktopHome";
+import { MobileHome } from "../mobile-home";
+import { AboutAlimentandoComAmor } from "../AboutAlimentandoComAmor";
+import { OngActions } from "../ongActions";
+
+const Home = () => {
+  return (
+    <>
+      <MobileHome />
+      <HomeDesktop />
+      <OngActions />
+      <AboutAlimentandoComAmor />
+    </>
+  );
+};
+
+export { Home };
