+++ src/components/Home/paragraph/types.ts
